refactor: FieldObject.js - fixed custom query input
